+++ C#/ProjetAutomate/Data/Models/Afpa_Lumiere.cs
@@ -8,7 +8,7 @@ namespace ProjetAutomate.Data.Models
     public partial class Afpa_Lumiere
     {
         public int IdLumiere { get; set; }
-        public int? ValeurLumiere { get; set; }
+        public float? ValeurLumiere { get; set; }
         public DateTime? DateLumiere { get; set; }
     }
 }
